chore: add useSingleSubmission hook - remove getString prop from SingleSubmissionWrite component
