+++ web/src/components/pages/contribution/sentence-collector/write/sentence-input-and-rules/sentence-input-and-rules.tsx
@@ -1,11 +1,10 @@
 import * as React from 'react'
-import { Localized } from '@fluent/react'
+import { Localized, useLocalization } from '@fluent/react'
 import classNames from 'classnames'
 
 import { EditIcon } from '../../../../../ui/icons'
 import { LabeledInput } from '../../../../../ui/ui'
 import { MultipleCombobox } from '../../../../../multiple-combobox'
-import { SingleSubmissionWriteProps } from '../single-submission-write/single-submission-write'
 import { Rules } from './rules'
 import { Instruction } from '../../instruction'
 import ExpandableInformation from '../../../../../expandable-information/expandable-information'
@@ -16,7 +15,6 @@ import URLS from '../../../../../../urls'
 import { useMultipleComboBox } from '../../../../../multiple-combobox/use-multiple-combox'
 
 type Props = {
-  getString: SingleSubmissionWriteProps['getString']
   handleSentenceInputChange: (
     event: React.ChangeEvent<HTMLTextAreaElement>
   ) => void
@@ -30,7 +28,6 @@ type Props = {
 }
 
 export const SentenceInputAndRules: React.FC<Props> = ({
-  getString,
   handleCitationChange,
   handleSentenceInputChange,
   sentenceDomains,
@@ -42,6 +39,8 @@ export const SentenceInputAndRules: React.FC<Props> = ({
 }) => {
   const isSentenceError = error && error !== SentenceSubmissionError.NO_CITATION
   const isCitationError = error === SentenceSubmissionError.NO_CITATION
+
+  const { l10n } = useLocalization()
 
   const { multipleComboBoxItems, inputValue, setInputValue } =
     useMultipleComboBox({
@@ -59,7 +58,7 @@ export const SentenceInputAndRules: React.FC<Props> = ({
         <div className="inputs">
           <Localized id="sentence" attrs={{ label: true }}>
             <LabeledTextArea
-              placeholder={getString('sentence-input-value')}
+              placeholder={l10n.getString('sentence-input-value')}
               className={classNames('sentence-input', {
                 'sentence-error': isSentenceError,
               })}
@@ -78,7 +77,7 @@ export const SentenceInputAndRules: React.FC<Props> = ({
           />
           <Localized id="citation" attrs={{ label: true }}>
             <LabeledInput
-              placeholder={getString('citation-input-value')}
+              placeholder={l10n.getString('citation-input-value')}
               className={classNames('citation-input', {
                 'citation-error': isCitationError,
               })}
