+++ web/src/components/pages/contribution/sentence-collector/write/single-submission-write/single-submission-write.tsx
@@ -1,150 +1,28 @@
-import {
-  Localized,
-  withLocalization,
-  WithLocalizationProps,
-} from '@fluent/react'
+import { Localized } from '@fluent/react'
 import * as React from 'react'
-import { useDispatch } from 'react-redux'
 
 import { QuestionIcon, SendIcon } from '../../../../../ui/icons'
 import { LabeledCheckbox, LinkButton } from '../../../../../ui/ui'
-import URLS from '../../../../../../urls'
 import { PrimaryButton } from '../../../../../primary-buttons/primary-buttons'
-import { useLocale } from '../../../../../locale-helpers'
 
 import { SentenceInputAndRules } from '../sentence-input-and-rules/sentence-input-and-rules'
-import { Sentences } from '../../../../../../stores/sentences'
-import {
-  sentenceDomains,
-  SentenceSubmission,
-  SentenceSubmissionError,
-} from 'common'
-import { Notifications } from '../../../../../../stores/notifications'
-import { useAction, useLanguages } from '../../../../../../hooks/store-hooks'
-import {
-  SingleSubmissionWriteActionType,
-  singleSubmissionWriteReducer,
-  SingleSubmissionWriteState,
-} from './single-submission-write.reducer'
-
-import { trackSingleSubmission } from '../../../../../../services/tracker'
+import { sentenceDomains } from 'common'
+import { useSingleSubmissionWrite } from './hooks/use-single-submission-write'
 
 import { COMMON_VOICE_EMAIL } from '../../../../../../constants'
+import URLS from '../../../../../../urls'
 
 import './single-submission-write.css'
 
-export type SingleSubmissionWriteProps = WithLocalizationProps
-
-const initialState: SingleSubmissionWriteState = {
-  sentence: '',
-  citation: '',
-  sentenceDomains: [],
-  error: undefined,
-  confirmPublicDomain: false,
-}
-
-const SingleSubmissionWrite: React.FC<SingleSubmissionWriteProps> = ({
-  getString,
-}) => {
-  const [state, singleSubmissionWriteDispatch] = React.useReducer(
-    singleSubmissionWriteReducer,
-    initialState
-  )
-
-  const [currentLocale] = useLocale()
-  const languages = useLanguages()
-
-  const localeId = languages.localeNameAndIDMapping.find(
-    locale => locale.name === currentLocale
-  ).id
-
-  const dispatch = useDispatch()
-
-  const createSentence = useAction(Sentences.actions.create)
-
-  const addNotification = ({
-    message,
-    type,
-  }: {
-    message: string
-    type: Notifications.NotificationType
-  }) => {
-    dispatch(Notifications.actions.addPill(message, type))
-  }
-
-  const handlePublicDomainChange = () => {
-    singleSubmissionWriteDispatch({
-      type: SingleSubmissionWriteActionType.SET_PUBLIC_DOMAIN,
-    })
-  }
-
-  const handleSentenceInputChange = (
-    event: React.ChangeEvent<HTMLTextAreaElement>
-  ) => {
-    singleSubmissionWriteDispatch({
-      type: SingleSubmissionWriteActionType.SET_SENTENCE,
-      payload: { sentence: event.target.value },
-    })
-  }
-
-  const handleCitationChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    singleSubmissionWriteDispatch({
-      type: SingleSubmissionWriteActionType.SET_CITATION,
-      payload: { citation: event.target.value },
-    })
-  }
-
-  const handleSentenceDomainChange = (domain: string[]) => {
-    singleSubmissionWriteDispatch({
-      type: SingleSubmissionWriteActionType.SET_SENTENCE_DOMAIN,
-      payload: { sentenceDomains: domain },
-    })
-  }
-
-  const handleSubmit = async (evt: React.SyntheticEvent) => {
-    evt.preventDefault()
-
-    const newSentence: SentenceSubmission = {
-      sentence: state.sentence,
-      source: state.citation,
-      localeId,
-      localeName: currentLocale,
-      domains: state.sentenceDomains,
-    }
-
-    try {
-      if (!state.citation) {
-        singleSubmissionWriteDispatch({
-          type: SingleSubmissionWriteActionType.ADD_SENTENCE_ERROR,
-          payload: { error: SentenceSubmissionError.NO_CITATION },
-        })
-      } else {
-        await createSentence(newSentence)
-
-        addNotification({
-          message: getString('add-sentence-success'),
-          type: 'success',
-        })
-
-        singleSubmissionWriteDispatch({
-          type: SingleSubmissionWriteActionType.ADD_SENTENCE_SUCCESS,
-        })
-
-        trackSingleSubmission('submit', currentLocale)
-      }
-    } catch (error) {
-      const errorMessage = JSON.parse(error.message)
-
-      singleSubmissionWriteDispatch({
-        type: SingleSubmissionWriteActionType.ADD_SENTENCE_ERROR,
-        payload: { error: errorMessage.errorType },
-      })
-      addNotification({
-        message: getString('add-sentence-error'),
-        type: 'error',
-      })
-    }
-  }
+const SingleSubmissionWrite: React.FC = () => {
+  const {
+    handleCitationChange,
+    handlePublicDomainChange,
+    handleSentenceDomainChange,
+    handleSentenceInputChange,
+    handleSubmit,
+    singleSentenceSubmissionState,
+  } = useSingleSubmissionWrite()
 
   return (
     <form
@@ -153,15 +31,16 @@ const SingleSubmissionWrite: React.FC<SingleSubmissionWriteProps> = ({
       onSubmit={handleSubmit}>
       <div className="inputs-and-rules-container">
         <SentenceInputAndRules
-          getString={getString}
           handleSentenceInputChange={handleSentenceInputChange}
           handleCitationChange={handleCitationChange}
-          sentence={state.sentence}
-          citation={state.citation}
+          sentence={singleSentenceSubmissionState.sentence}
+          citation={singleSentenceSubmissionState.citation}
           sentenceDomains={sentenceDomains}
           setSelectedSentenceDomains={handleSentenceDomainChange}
-          selectedSentenceDomains={state.sentenceDomains}
-          error={state.error}
+          selectedSentenceDomains={
+            singleSentenceSubmissionState.sentenceDomains
+          }
+          error={singleSentenceSubmissionState.error}
         />
       </div>
 
@@ -206,8 +85,8 @@ const SingleSubmissionWrite: React.FC<SingleSubmissionWriteProps> = ({
                 <span />
               </Localized>
             }
-            disabled={state.sentence.length === 0}
-            checked={state.confirmPublicDomain}
+            disabled={singleSentenceSubmissionState.sentence.length === 0}
+            checked={singleSentenceSubmissionState.confirmPublicDomain}
             required
             onChange={handlePublicDomainChange}
             data-testid="public-domain-checkbox"
@@ -216,7 +95,7 @@ const SingleSubmissionWrite: React.FC<SingleSubmissionWriteProps> = ({
             <PrimaryButton
               className="submit"
               type="submit"
-              disabled={!state.confirmPublicDomain}
+              disabled={!singleSentenceSubmissionState.confirmPublicDomain}
               data-testid="submit-button"
             />
           </Localized>
@@ -226,4 +105,4 @@ const SingleSubmissionWrite: React.FC<SingleSubmissionWriteProps> = ({
   )
 }
 
-export default withLocalization(SingleSubmissionWrite)
+export default SingleSubmissionWrite
